requirements.txt
-e .
django~=3.2.9
djangorestframework~=3.12.4
tox~=3.24.4
pytest-django~=4.4.0
isort~=5.10.1
flake8-docstrings~=1.6.0
flake8~=4.0.1
black~=21.11b1
model-bakery~=1.3.3
pytest-cov~=3.0.0